Added database translation for LBGs (experimental)
An experiment at translating local BEST group names from the default English,
to Slovak, for displaying on the SVG map, via a table of Slovak names that
can be JOINed to the original table via the 2-letter ID used on the original
webpage best.eu.org (NOTE: It is not a foreign key of the translation table,
as the enforcement of referential integrity would make updating and refreshing
the original table problematic, it is merely a primary key for the translation
table).

Also added a separate tab in BEST DB settings with the update history table
and a button to refresh the translation tables in a database (theoretically,
it should work with multiple languages and thus multiple translation tables,
as clicking on the button should just refresh each of the tables one by one,
but we have not yet tested it). Currently, Slovak names of LBGs are retrieved
from an XML file in the assets directory.

To test these features just try to change entries in the XML file (lbgs_sk.xml)
and watch it reflected on the map on the page. If you want to test functionality
with multiple languages, the array of language codes is a static member variable
of the database class (not sure honestly, where else to put it). Just add a language
code in the array, create another XML file, refresh the database and change the select
statement in lbgs_clickable_map.php.

PS: Also renamed id of the <div> wrapper of the SVG map to "lbg_map" and sorted LBG entries
in the map-optimized.svg file by their id attributes in reverse alphabetical order (Z-A).
Not sure it optimizes eny parsing, but nice to have. :)